+++ app/server/models/transfer_card.py
@@ -6,7 +6,7 @@ from sqlalchemy import desc
 from server import db
 from server.models.utils import ModelBase
 import server.models.credit_transfer
-from server.models.transfer_card_usage import TransferCardUsage
+from server.models.transfer_card_state import TransferCardState
 from server.utils.transfer_enums import TransferTypeEnum, TransferStatusEnum, TransferSubTypeEnum
 from server.utils.audit_history import manually_add_history_entry
 from server.exceptions import NoTransferCardError
@@ -36,7 +36,7 @@ class TransferCard(ModelBase):
         foreign_keys='CreditTransfer.sender_transfer_card_id'
     )
 
-    last_usages = db.relationship(TransferCardUsage, order_by=desc(TransferCardUsage.session_number), lazy='dynamic')
+    last_usages = db.relationship(TransferCardState, order_by=desc(TransferCardState.session_number), lazy='dynamic')
 
     def disable(self):
         # disabling cards is permanent, hence no setter to enable a disabled card
