+++ app/test_app/functional/me_api/test_me_credit_transfer.py
@@ -98,7 +98,7 @@ def test_create_me_credit_transfer_api(
     # Make sure the transfer card usage was stored and all the joins work correctly  
     card_usage_object = credit_transfer.transfer_card_state 
     assert card_usage_object.transfer_card.public_serial_number == '222222'
-    assert card_usage_object.transfer_card.transfer_card_usages == [card_usage_object]
+    assert card_usage_object.transfer_card.transfer_card_states == [card_usage_object]
     assert card_usage_object.session_number == 5
     assert card_usage_object.amount_deducted == 100
     assert card_usage_object.amount_loaded == 150
